Refactoring: Use unique_ptr in lists and cleanup unique_ptr declarations
This is the final step in making the code elegantly exception-safe. Now
that all objects are stored in either unique_ptr's or containers of
unique_ptr's, an exception can occur at any time and objects will still be
properly destructed. Accidental memory leaks are also much more difficult
to cause.