Bugfix in determining Skype call duration
It turns out that the timestamp difference between call events is not always reliable. Sometimes the call start and end events will both be emitted with the timestamp of the end event, leading to an incorrect interpretation of the call as having failed. With this fix in place, the true duration will be read from the param_value column, if available.