Created abstraction for the files from which convos are extracted
This is part of adding support for extracting conversations from entries in
a ZIP archive, etc. Such an entry is a sort of “pseudo-file” in that it does not
exist on the disk under a regular directory entry, but it has a logical filename,
attributes, and content that can be extracted.