Fix destructors for "true interface" classes
Classes like ISerializable, IPolymorphic etc. (whose names have the "I" prefix) are meant to be used only to force (and signal) the presence of certain methods in their parent classes. An object will never be owned (i.e. put in a vector, returned) through a pointer to a Ixxx base class. To signal this, we make the destructors in such interfaces protected, which prevents accidental use as a base class for storage.